Embedded solvers merge
