+++ applications/FluidDynamicsApplication/python_scripts/python_solvers_wrapper_fluid.py
@@ -14,11 +14,8 @@ def CreateSolverByParameters(model, solver_settings, parallelism):
         elif (solver_type == "FractionalStep"):
             solver_module_name = "navier_stokes_solver_fractionalstep"
 
-        elif ((solver_type == "Embedded") or (solver_type == "EmbeddedDevelopment") or (solver_type == "EmbeddedAusasDevelopment")):
+        elif ((solver_type == "Embedded") or (solver_type == "EmbeddedDevelopment") or (solver_type == "EmbeddedAusas") or (solver_type == "EmbeddedAusasDevelopment")):
             solver_module_name = "navier_stokes_embedded_solver"
-
-        elif (solver_type == "EmbeddedAusas"):
-            solver_module_name = "navier_stokes_embedded_ausas_solver"
 
         elif (solver_type == "Compressible"):
             solver_module_name = "navier_stokes_compressible_solver"
@@ -37,11 +34,8 @@ def CreateSolverByParameters(model, solver_settings, parallelism):
         elif (solver_type == "FractionalStep"):
             solver_module_name = "trilinos_navier_stokes_solver_fractionalstep"
 
-        elif (solver_type == "Embedded"):
+        elif ((solver_type == "Embedded") or (solver_type == "EmbeddedDevelopment") or (solver_type == "EmbeddedAusas") or (solver_type == "EmbeddedAusasDevelopment")):
             solver_module_name = "trilinos_navier_stokes_embedded_solver"
-
-        elif (solver_type == "EmbeddedAusas"):
-            solver_module_name = "trilinos_navier_stokes_embedded_ausas_solver"
 
         else:
             raise Exception("the requested solver type is not in the python solvers wrapper. Solver type is : " + solver_type)
@@ -70,5 +64,8 @@ def CreateSolver(model, custom_settings):
         from KratosMultiphysics import MeshMovingApplication
         import ale_fluid_solver
         return ale_fluid_solver.CreateSolver(model, solver_settings, parallelism)
+
+    if solver_settings.Has("fm_ale_settings"):
+        KratosMultiphysics.CheckRegisteredApplications("MeshMovingApplication")
 
     return CreateSolverByParameters(model, solver_settings, parallelism)
